requirements.txt
biopython
regex
inquirer
readchar
